Move NodeCPUBurstCfg defination to slo-controller/config/config.go
