--- apis/extension/slo-config.go
@@ -1,34 +1,0 @@
-/*
-Copyright 2022 The Koordinator Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package extension
-
-import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	slov1alpha1 "github.com/koordinator-sh/koordinator/apis/slo/v1alpha1"
-)
-
-type NodeCPUBurstCfg struct {
-	// an empty label selector matches all objects while a nil label selector matches no objects
-	NodeSelector *metav1.LabelSelector `json:"nodeSelector,omitempty"`
-	*slov1alpha1.CPUBurstStrategy
-}
-
-type CPUBurstCfg struct {
-	ClusterStrategy *slov1alpha1.CPUBurstStrategy `json:"clusterStrategy,omitempty"`
-	NodeStrategies  []NodeCPUBurstCfg             `json:"nodeStrategies,omitempty"`
-}
